requirements.txt
ultralytics==8.3.241
torch==2.9.1+cpu
torchvision==0.24.1
opencv-python
pillow
streamlit
numpy
pyyaml
lxml
tqdm
